resume_enhancer/requirements.txt
python-dotenv~=1.0.1
langchain~=0.1.5
requests~=2.31.0
streamlit~=1.31.0
openai~=1.51.0
pypdf~=4.0.2
streamlit_chat~=0.1.1
pymongo~=4.6.2
tavily-python==0.3.1
attrs~=23.2.0
distro~=1.9.0
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.3
pathlib==1.0.1
configparser==6.0.1
filelock==3.13.1
PyMuPDF==1.24.1
langchain-openai~=0.1.3
langchainhub~=0.1.15
python_docx==1.0.1
langchain-core~=0.1.43
PyPDF2~=3.0.1
pdfplumber==0.5.28
pytesseract==0.3.10
fpdf==1.7.2
pdf2image==1.16.0
fpdf2==2.5.6  # or the latest version available
